Some structural changes and add more tests
