+++ tests/FINDOLOGIC/Api/Tests/FindologicClientTest.php
