+++ tests/FINDOLOGIC/Api/Tests/FindologicConfigTest.php
